# Api/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "APP API",
    "description": "",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/signup-update": {
      "post": {
        "tags": [
          "Accounts"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Add new user.",
            "schema": {
              "$ref": "#/definitions/Accounts"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sign up success",
            "schema": {
              "type": "boolean",
              "example": true,
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/login": {
      "post": {
        "tags": [
          "Accounts"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Login.",
            "schema": {
              "$ref": "#/definitions/Login"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Login success",
            "schema": {
              "$ref": "#/definitions/Accounts"
            }
          },
          "500": {
            "description": "Login failt",
            "schema": {
              "type": "string",
              "example": "error",
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/get-list": {
      "get": {
        "tags": [
          "Accounts"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "Get list success",
            "schema": {
              "$ref": "#/definitions/Accounts"
            }
          },
          "500": {
            "description": "Get fail",
            "schema": {
              "type": "string",
              "example": "error",
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/get-by-id/{_id}": {
      "get": {
        "tags": [
          "Accounts"
        ],
        "description": "",
        "parameters": [
          {
            "name": "_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Get success",
            "schema": {
              "$ref": "#/definitions/Accounts"
            }
          },
          "500": {
            "description": "Get fail",
            "schema": {
              "type": "string",
              "example": "error",
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/delete/{_id}": {
      "delete": {
        "tags": [
          "Accounts"
        ],
        "description": "",
        "parameters": [
          {
            "name": "_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Delete success",
            "schema": {
              "type": "boolean",
              "example": true,
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Delete fail",
            "schema": {
              "type": "string",
              "example": "error",
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/get-by-email/{email}": {
      "get": {
        "tags": [
          "Accounts"
        ],
        "description": "",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Get success",
            "schema": {
              "$ref": "#/definitions/Accounts"
            }
          },
          "500": {
            "description": "Get fail",
            "schema": {
              "type": "string",
              "example": "error",
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "Accounts": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "string"
        },
        "email": {
          "type": "string",
          "example": "string"
        },
        "phone": {
          "type": "string",
          "example": "string"
        },
        "password": {
          "type": "string",
          "example": "string"
        }
      },
      "required": [
        "name",
        "email",
        "phone",
        "password"
      ]
    },
    "Login": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "example": "string"
        },
        "password": {
          "type": "string",
          "example": "string"
        }
      }
    },
...
  }
}